Move input remapping code to configuration.c
